requirements.txt
PyQt6>=6.5.0
pyserial>=3.5
python-dotenv>=1.0.0
